+++ etc/block-components/enter-numbers/ui/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @milaboratories/milaboratories.test-enter-numbers.ui
 
+## 1.0.121
+
+### Patch Changes
+
+- Updated dependencies [5884403]
+  - @platforma-sdk/ui-vue@1.14.2
+
+## 1.0.120
+
+### Patch Changes
+
+- Updated dependencies [a886400]
+  - @platforma-sdk/ui-vue@1.14.1
+  - @milaboratories/milaboratories.test-enter-numbers.model@1.0.29
+
+## 1.0.119
+
+### Patch Changes
+
+- Updated dependencies [dfad2d0]
+  - @platforma-sdk/ui-vue@1.14.0
+  - @milaboratories/milaboratories.test-enter-numbers.model@1.0.28
+
+## 1.0.118
+
+### Patch Changes
+
+- Updated dependencies [a85e3f7]
+  - @platforma-sdk/ui-vue@1.13.18
+
+## 1.0.117
+
+### Patch Changes
+
+- @platforma-sdk/ui-vue@1.13.17
+
 ## 1.0.116
 
 ### Patch Changes