+++ etc/block-components/ui-examples/model/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/milaboratories.ui-examples.model
 
+## 1.0.45
+
+### Patch Changes
+
+- Updated dependencies [a886400]
+  - @platforma-sdk/model@1.14.1
+
+## 1.0.44
+
+### Patch Changes
+
+- @platforma-sdk/model@1.14.0
+
 ## 1.0.43
 
 ### Patch Changes