+++ etc/blocks/enter-numbers/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @milaboratories/milaboratories.test-enter-numbers
 
+## 1.0.121
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.121
+
+## 1.0.120
+
+### Patch Changes
+
+- Updated dependencies [a886400]
+  - @platforma-sdk/model@1.14.1
+  - @milaboratories/milaboratories.test-enter-numbers.ui@1.0.120
+  - @milaboratories/milaboratories.test-enter-numbers.model@1.0.29
+
+## 1.0.119
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.workflow@1.0.3
+- @platforma-sdk/model@1.14.0
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.119
+- @milaboratories/milaboratories.test-enter-numbers.model@1.0.28
+
+## 1.0.118
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.118
+
+## 1.0.117
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.117
+
 ## 1.0.116
 
 ### Patch Changes