+++ etc/blocks/model-test/model/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/milaboratories.test-block-model.model
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [a886400]
+  - @platforma-sdk/model@1.14.1
+
+## 1.0.10
+
+### Patch Changes
+
+- @platforma-sdk/model@1.14.0
+
 ## 1.0.9
 
 ### Patch Changes